Added the possibility to read a topology.
Not quite finished since the age and the error of the different
components are not currently saved in the output. The current
implementation allows to read a topology from a json file, and init the
initial graph with the topology. If no topology is provided, the graph
is initialized randomly.